다른 서비스를 지원하기 위해 네이버 뉴스 처리를 별도 스크립트로 분리
- 서비스에 따라 스크립트를 분리하기 위해 동적 로딩 적용
